chore: Fix story
